+++ backend/windmill-common/src/global_settings.rs
@@ -4,6 +4,7 @@ pub const BASE_URL_SETTING: &str = "base_url";
 pub const OAUTH_SETTING: &str = "oauths";
 pub const RETENTION_PERIOD_SECS_SETTING: &str = "retention_period_secs";
 pub const REQUEST_SIZE_LIMIT_SETTING: &str = "request_size_limit_mb";
+pub const LICENSE_KEY_SETTING: &str = "license_key";
 
 pub const ENV_SETTINGS: [&str; 54] = [
     "DISABLE_NSJAIL",