+++ backend/windmill-common/src/worker.rs
@@ -78,15 +78,12 @@ pub async fn reload_custom_tags_setting(db: &DB) -> error::Result<()> {
 
     let custom_tags = process_custom_tags(tags);
 
-    {
-        let l = CUSTOM_TAGS_PER_WORKSPACE.read().await;
-        if l.clone() == custom_tags {
-            tracing::info!("Custom tags setting unchanged, skipping update");
-            return Ok(());
-        } else {
-            tracing::info!("Custom tags setting changed, updating");
-        }
-    }
+    tracing::info!(
+        "Loaded setting custom tags, common: {:?}, per-workspace: {:?}",
+        custom_tags.0,
+        custom_tags.1,
+    );
+
     {
         let mut l = CUSTOM_TAGS_PER_WORKSPACE.write().await;
         *l = custom_tags.clone()